Split sidebar, buttons to separate modules
